+++ NOTES
@@ -1,0 +1,5 @@
+https://pypi.org/project/torch/#files
+
+2.5.1 e.g.
+
+Only wheels, no sdists.
